+++ src/SideMenu/core/Promotion.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classnames from 'classnames';
 import styles from './styles.scss';
 
-const Promotion = ({children}) =>
-  <div className={styles.promotion} data-hook="menu-promotion">
+const Promotion = ({children, className}) =>
+  <div className={classnames(styles.promotion, className)} data-hook="menu-promotion">
     {children}
   </div>;
 
 Promotion.propTypes = {
+  className: PropTypes.string,
   children: PropTypes.node
 };
 
 export default Promotion;
-