+++ src/SideMenu/core/navigation/BackLink.js
@@ -1,18 +1,24 @@
 import React from 'react';
-import {func, node} from 'prop-types';
+import PropTypes from 'prop-types';
+import classnames from 'classnames';
 import ChevronLeft from 'wix-ui-icons-common/ChevronLeft';
 import styles from './styles.scss';
 
-const BackLink = ({onBackHandler, children}) => (
-  <a className={styles.backLink} onClick={onBackHandler} data-hook="menu-navigation-back-link">
+const BackLink = ({onBackHandler, className, children}) => (
+  <a
+    className={classnames(styles.backLink, className)}
+    onClick={onBackHandler}
+    data-hook="menu-navigation-back-link"
+    >
     <span className={styles.backArrow}><ChevronLeft size="14px"/></span>
     <span className={styles.backLabel}>{children}</span>
   </a>
 );
 
 BackLink.propTypes = {
-  onBackHandler: func,
-  children: node
+  onBackHandler: PropTypes.func,
+  className: PropTypes.string,
+  children: PropTypes.node
 };
 
 export default BackLink;