+++ src/SideMenu/core/navigation/Link.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {bool, node, string} from 'prop-types';
+import PropTypes from 'prop-types';
 import ChevronRight from 'wix-ui-icons-common/ChevronRight';
 import LinkLayout from './LinkLayout';
 import styles from './styles.scss';
@@ -20,13 +20,13 @@ Link.defaultProps = {
 };
 
 Link.propTypes = {
-  children: node,
-  isActive: bool,
-  withArrow: bool,
-  badge: node,
-  isDiminishedHover: bool,
-  dataHook: string,
-  disabled: bool
+  children: PropTypes.node,
+  isActive: PropTypes.bool,
+  withArrow: PropTypes.bool,
+  badge: PropTypes.node,
+  isDiminishedHover: PropTypes.bool,
+  dataHook: PropTypes.string,
+  disabled: PropTypes.bool
 };
 
 export default Link;