+++ src/SideMenu/core/navigation/LinkLayout.js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import styles from './styles.scss';
 
-const LinkLayout = ({children, isDiminishedHover, isActive, disabled, ...rest}) =>
+const LinkLayout = ({children, isDiminishedHover, isActive, disabled, className, ...rest}) =>
   <span
     className={classnames(styles.linkLayout, {
       [styles.linkActive]: isActive,
       [styles.linkDiminishedHover]: isDiminishedHover,
       [styles.linkDisabled]: disabled
-    })}
+    }, className)}
     data-hook="menu-navigation-link-wrapper"
     {...rest}
     >
@@ -20,7 +20,8 @@ LinkLayout.propTypes = {
   children: PropTypes.node,
   isActive: PropTypes.bool,
   isDiminishedHover: PropTypes.bool,
-  disabled: PropTypes.bool
+  disabled: PropTypes.bool,
+  className: PropTypes.string
 };
 
 export default LinkLayout;