+++ src/SideMenu/core/navigation/Separator.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classnames from 'classnames';
 import styles from './styles.scss';
 
-const Separator = ({children}) =>
-  <div className={styles.separator} data-hook="menu-navigation-separator">
+const Separator = ({children, className}) =>
+  <div className={classnames(styles.separator, className)} data-hook="menu-navigation-separator">
     {children}
   </div>;
 
 Separator.propTypes = {
+  className: PropTypes.string,
   children: PropTypes.node
 };
 
 export default Separator;
-