+++ src/SideMenu/core/navigation/index.js
@@ -1,13 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classnames from 'classnames';
 import styles from './styles.scss';
 
-const Navigation = ({children}) =>
-  <div className={styles.navigation} data-hook="menu-navigation">
+const Navigation = ({children, className}) =>
+  <div className={classnames(styles.navigation, className)} data-hook="menu-navigation">
     {children}
   </div>;
 
 Navigation.propTypes = {
+  className: PropTypes.string,
   children: PropTypes.node
 };